# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive WhatsApp Business API built with Go, following Clean Architecture principles.\nProvides endpoints for session management, messaging, contacts, groups, media handling, and integrations.",
        "title": "zpwoot WhatsApp API",
        "contact": {
            "name": "zpwoot Support",
            "url": "https://github.com/your-org/zpwoot",
            "email": "support@zpwoot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sessions/{sessionId}/proxy/find": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get proxy configuration for a specific session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Find Proxy Configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProxyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Configuration not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{sessionId}/proxy/set": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create or update proxy configuration for a session (upsert operation). To disable/remove proxy, set enabled: false",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Set Proxy Configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Proxy configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProxyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.ProxyResponse"
                        }
                    },
                    "201": {
                        "description": "Configuration created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.ProxyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "type": "string",
                    "example": "name is required"
                }
            }
        },
...
        "dto.CreateProxyRequest": {
            "type": "object",
            "required": [
                "host",
                "port",
                "protocol"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string",
                    "minLength": 1
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1
                },
                "protocol": {
                    "type": "string",
                    "enum": [
                        "http",
                        "https",
                        "socks4",
                        "socks5"
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.ProxyResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication. Use the value from .env file (API_KEY variable). Send as 'Authorization: your-api-key' (without Bearer prefix)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}